triangulation is happening but not sure if correct
