+++ dev/dev_real_time_triangulator_bigger_numba.py
@@ -33,18 +33,18 @@ config = Configurator(session_path)
 charuco: Charuco = config.get_charuco()
 camera_array: CameraArray = config.get_camera_array()
 
-# logger.info(f"Creating RecordedStreamPool")
-# stream_pool = RecordedStreamPool(session_path, charuco=charuco)
-# logger.info("Creating Synchronizer")
-# syncr = Synchronizer(stream_pool.streams, fps_target=None)
-# stream_pool.play_videos()
+logger.info(f"Creating RecordedStreamPool")
+stream_pool = RecordedStreamPool(session_path, charuco=charuco)
+logger.info("Creating Synchronizer")
+syncr = Synchronizer(stream_pool.streams, fps_target=None)
+stream_pool.play_videos()
 
 
 #### Basic code for interfacing with in-progress RealTimeTriangulator
 #### Just run off of saved point_data.csv for development/testing
-# real_time_triangulator = RealTimeTriangulator(camera_array, syncr)
-# while real_time_triangulator.running:
-#     sleep(1)
+real_time_triangulator = RealTimeTriangulator(camera_array, syncr)
+while real_time_triangulator.running:
+    sleep(1)
 
 # # packet 20 appears to be a good sample for development...
 # sync_packet: SyncPacket = real_time_triangulator._sync_packet_history[20]
@@ -56,110 +56,62 @@ camera_array: CameraArray = config.get_camera_array()
 # save out point packets to csv for storage and refresh later
 
 # Load point data
-points_xy = pd.read_csv(Path(session_path, "point_data.csv"))
+# points_xy = pd.read_csv(Path(session_path, "point_data.csv"))
 
-camera_indices = points_xy["port"].to_numpy()
-sync_indices = points_xy["sync_index"].to_numpy()
-point_indices = points_xy["point_id"].to_numpy()
-img_x = points_xy["img_loc_x"].to_numpy()
-img_y = points_xy["img_loc_y"].to_numpy()
-img = np.vstack([img_x, img_y]).T
-
-
-@jit(nopython=True)
-def unique_with_counts(arr):
-    sorted_arr = np.sort(arr)
-    unique_values = [sorted_arr[0]]
-    counts = [1]
-
-    for i in range(1, len(sorted_arr)):
-        if sorted_arr[i] != sorted_arr[i - 1]:
-            unique_values.append(sorted_arr[i])
-            counts.append(1)
-        else:
-            counts[-1] += 1
-
-    return np.array(unique_values), np.array(counts)
-####################### Function from ChatGPT Discussion #################################
-# def triangulate_points_modified(point_packets: Dict[int, PointPacket], camera_data: Dict[int, CameraData]) -> Dict[int, np.ndarray]:
+# camera_indices = points_xy["port"].to_numpy()
+# sync_indices = points_xy["sync_index"].to_numpy()
+# point_indices = points_xy["point_id"].to_numpy()
+# img_x = points_xy["img_loc_x"].to_numpy()
+# img_y = points_xy["img_loc_y"].to_numpy()
+# img = np.vstack([img_x, img_y]).T
 
 
-@jit(nopython=True, parallel=True, cache=True)
-def triangulate_sync_index(
-    projection_matrices, current_camera_indices, current_point_id, current_img
-):
-    # sync_indices_xyz = List()
-    point_indices_xyz = List()
-    obj_xyz = List()
-
-    unique_points, point_counts = unique_with_counts(current_point_id)
-    for index in range(len(point_counts)):
-        if point_counts[index] > 1:
-            # triangulate that points...
-            point = unique_points[index]
-            points_xy = current_img[current_point_id == point]
-            camera_ids = current_camera_indices[current_point_id == point]
-            # logger.info(f"Calculating xyz for point {point} at sync index {sync_id}")
-            # point_xyz = triangulate_simple(points_xy, camera_ids, projection_matrices)
-
-            num_cams = len(camera_ids)
-            A = np.zeros((num_cams * 2, 4))
-            for i in range(num_cams):
-                x, y = points_xy[i]
-                P = projection_matrices[camera_ids[i]]
-                A[(i * 2) : (i * 2 + 1)] = x * P[2] - P[0]
-                A[(i * 2 + 1) : (i * 2 + 2)] = y * P[2] - P[1]
-            u, s, vh = np.linalg.svd(A, full_matrices=True)
-            point_xyzw = vh[-1]
-            point_xyz = point_xyzw[:3] / point_xyzw[3]
-
-            # sync_indices_xyz.append(sync_id)
-            point_indices_xyz.append(point)
-            obj_xyz.append(point_xyz)
-
-    return point_indices_xyz, obj_xyz
+# ####################### Function from ChatGPT Discussion #################################
+# # def triangulate_points_modified(point_packets: Dict[int, PointPacket], camera_data: Dict[int, CameraData]) -> Dict[int, np.ndarray]:
 
 
-# initialize numba specific datastructures that will be used to access data within function
 
-projection_matrices = Dict()
-# projection_matrices = {}
-for port, cam in camera_array.cameras.items():
-    projection_matrices[port] = cam.projection_matrix
 
-# all_sync_indices_xyz = List()
-# all_point_indices_xyz = List()
-# all_obj_xyz = List()
+# # initialize numba specific datastructures that will be used to access data within function
 
-all_sync_id_xyz = []
-all_point_id_xyz = []
-all_points_xyz = []
+# projection_matrices = Dict()
+# # projection_matrices = {}
+# for port, cam in camera_array.cameras.items():
+#     projection_matrices[port] = cam.projection_matrix
 
-logger.info(f"Begin processing sync indices")
-for sync_index in np.unique(sync_indices):
-    # these arrays should be providing the basic structure of the PointPacket
-    cameras_at_sync_index = camera_indices[sync_indices == sync_index]
-    points_at_sync_index = point_indices[sync_indices == sync_index]
-    img_xy_at_sync_index = img[sync_indices == sync_index]
+# # all_sync_indices_xyz = List()
+# # all_point_indices_xyz = List()
+# # all_obj_xyz = List()
 
-    start_sync_index = time()
-    # only attempt to process points with multiple views
-    # iterated across the current points to find those with multiple views
+# all_sync_id_xyz = []
+# all_point_id_xyz = []
+# all_points_xyz = []
 
-    point_id_xyz, points_xyz = triangulate_sync_index(
-        projection_matrices, cameras_at_sync_index, points_at_sync_index, img_xy_at_sync_index
-    )
-    stop_sync_index = time()
-    current_sync_index_list = [sync_index]*len(point_id_xyz)
-    all_sync_id_xyz.extend(current_sync_index_list)
-    all_point_id_xyz.extend(point_id_xyz)
-    all_points_xyz.extend(points_xyz)
+# logger.info(f"Begin processing sync indices")
+# for sync_index in np.unique(sync_indices):
+#     # these arrays should be providing the basic structure of the PointPacket
+#     cameras_at_sync_index = camera_indices[sync_indices == sync_index]
+#     points_at_sync_index = point_indices[sync_indices == sync_index]
+#     img_xy = img[sync_indices == sync_index]
 
-    # logger.info(f"Finished processing sync id {sync_id} at {stop_sync_index}")
-    elapsed_time = start_sync_index - stop_sync_index
-    logger.info(f"Elapsed time to process sync index {sync_index} is {elapsed_time}")
+#     start_sync_index = time()
+#     # only attempt to process points with multiple views
+#     # iterated across the current points to find those with multiple views
 
-    # only attempt to process points with multiple views
+#     point_id_xyz, points_xyz = triangulate_sync_index(
+#         projection_matrices, cameras_at_sync_index, points_at_sync_index, img_xy
+#     )
+#     stop_sync_index = time()
+#     current_sync_index_list = [sync_index]*len(point_id_xyz)
+#     all_sync_id_xyz.extend(current_sync_index_list)
+#     all_point_id_xyz.extend(point_id_xyz)
+#     all_points_xyz.extend(points_xyz)
+
+#     # logger.info(f"Finished processing sync id {sync_id} at {stop_sync_index}")
+#     elapsed_time = start_sync_index - stop_sync_index
+#     logger.info(f"Elapsed time to process sync index {sync_index} is {elapsed_time}")
+
+#     # only attempt to process points with multiple views
 
 # %%
